Add configuration files and update imports for insect and flower tracking
