+++ src/insect_recorder.py
@@ -1,6 +1,5 @@
 import cv2
 import numpy as np
-import pandas as pd
 import os
 import logging
 import matplotlib.pyplot as plt
